Project has been refactored.

+++ src/main/java/br/com/alansep/dailyhelper/model/Task.java
@@ -1,6 +1,5 @@
 package br.com.alansep.dailyhelper.model;
 
-
 import br.com.alansep.dailyhelper.model.enums.Period;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -12,21 +11,19 @@ import lombok.Setter;
 @Builder
 @EqualsAndHashCode(of = "id")
 public class Task {
-	
-	private Byte id;
+
+	private Integer id;
 	private String title;
 	private String description;
 	private Period period;
-	
+
 	@Override
 	public String toString() {
-		if(id == null) {
+		if (id == null) {
 			return title;
 		} else {
-			return (id + 1) + " - " + title;			
+			return (id + 1) + " - " + title;
 		}
 	}
-	
-	
 
 }
